Combine forward destinations and query types charts into one component
Named GenericDoughnutChart.
